Improved payment-method view

+++ db/migrate/20190730033800_create_t_metodo_pagos.rb
@@ -3,8 +3,8 @@ class CreateTMetodoPagos < ActiveRecord::Migration[5.2]
     create_table :t_metodo_pagos do |t|
       t.string :forma_pago, null: false
       t.string :descripcion, null: false
-      t.float :minimo
-      t.float :maximo  
+      t.decimal :minimo
+      t.decimal :maximo  
       t.integer :estatus, null: false
 
       t.timestamps
